add swagger for docs, fix project

+++ routes/activities.js
@@ -1,11 +1,11 @@
 const router = require("express").Router();
 const Activity = require("../models/Activity");
-const Album = require("../models/Album");
+const User = require("../models/User");
 
 //create activity
 router.post("/", async (req, res) => {
-  const newActivity = new Activity(req.body);
   try {
+    const newActivity = new Activity(req.body);
     const savedActivity = await newActivity.save();
     res.status(200).json(savedActivity);
   } catch (err) {
@@ -15,13 +15,16 @@ router.post("/", async (req, res) => {
 
 //update activity
 router.put("/:id", async (req, res) => {
-  const activity = await Activity.findById(req.params.id);
   try {
+    const activity = await Activity.findById(req.params.id);
+    if (!activity) {
+      return res.status(404).json("Activity not found");
+    }
     if (activity.userId === req.body.userId) {
       await activity.updateOne({ $set: req.body });
       res.status(200).json("Activity updated!");
     } else {
-      res.status(500).json("You can only update your activity");
+      res.status(403).json("You can only update your activity");
     }
   } catch (err) {
     res.status(500).json(err);
@@ -30,15 +33,18 @@ router.put("/:id", async (req, res) => {
 
 //delete activity
 router.delete("/:id", async (req, res) => {
-  const activity = await Activity.findById(req.params.id);
   try {
+    const activity = await Activity.findById(req.params.id);
+    if (!activity) {
+      return res.status(404).json("Activity not found");
+    }
     if (activity.userId === req.body.userId) {
       // await activity.deleteOne({ $set: req.body });
       await Activity.findByIdAndDelete(req.params.id);
 
       res.status(200).json("Activity deleted!");
     } else {
-      res.status(500).json("You can only delete your activity");
+      res.status(403).json("You can only delete your activity");
     }
   } catch (err) {
     res.status(500).json(err);
@@ -49,17 +55,10 @@ router.delete("/:id", async (req, res) => {
 router.get("/:id", async (req, res) => {
   try {
     const activity = await Activity.findById(req.params.id);
+    if (!activity) {
+      return res.status(404).json("Activity not found");
+    }
     res.status(200).json(activity);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-//get all activity
-router.get("/timeline/all", async (req, res) => {
-  try {
-    const userActivity = await Activity.find().sort({ createdAt: "desc" });
-    res.json(userActivity);
   } catch (err) {
     res.status(500).json(err);
   }
@@ -68,6 +67,10 @@ router.get("/timeline/all", async (req, res) => {
 //get all activity by id user
 router.get("/all/:id", async (req, res) => {
   try {
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      return res.status(404).json("User not found!");
+    }
     const userActivity = await Activity.find({ userId: req.params.id }).sort({
       createdAt: "desc",
     });
@@ -80,6 +83,10 @@ router.get("/all/:id", async (req, res) => {
 // get last activity by id user
 router.get("/last/:id", async (req, res) => {
   try {
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      return res.status(404).json("User not found!");
+    }
     const userActivity = await Activity.findOne({ userId: req.params.id }).sort(
       {
         startDate: "desc",
@@ -92,7 +99,7 @@ router.get("/last/:id", async (req, res) => {
 });
 
 // search activities by query
-router.get("/search/:search/", async (req, res) => {
+router.get("/search/:search", async (req, res) => {
   try {
     var activities = await Activity.find({
       $or: [
