+++ routes/posts.js
@@ -4,8 +4,8 @@ const User = require("../models/User");
 
 //create post
 router.post("/", async (req, res) => {
-  const newPost = new Post(req.body);
   try {
+    const newPost = new Post(req.body);
     const savedPost = await newPost.save();
     res.status(200).json(savedPost);
   } catch (err) {
@@ -15,13 +15,16 @@ router.post("/", async (req, res) => {
 
 //update post
 router.put("/:id", async (req, res) => {
-  const post = await Post.findById(req.params.id);
   try {
+    const post = await Post.findById(req.params.id);
+    if (!post) {
+      return res.status(404).json("Post not found");
+    }
     if (post.userId === req.body.userId) {
       await post.updateOne({ $set: req.body });
       res.status(200).json("Post updated!");
     } else {
-      res.status(500).json("You can only update your post");
+      res.status(403).json("You can only update your post");
     }
   } catch (err) {
     res.status(500).json(err);
@@ -30,15 +33,18 @@ router.put("/:id", async (req, res) => {
 
 //delete post
 router.delete("/:id", async (req, res) => {
-  const post = await Post.findById(req.params.id);
   try {
+    const post = await Post.findById(req.params.id);
+    if (!post) {
+      return res.status(404).json("Post not found");
+    }
     if (post.userId === req.body.userId) {
       // await post.deleteOne({ $set: req.body });
       await Post.findByIdAndDelete(req.params.id);
 
       res.status(200).json("Post deleted!");
     } else {
-      res.status(500).json("You can only delete your post");
+      res.status(403).json("You can only delete your post");
     }
   } catch (err) {
     res.status(500).json(err);
@@ -49,6 +55,9 @@ router.delete("/:id", async (req, res) => {
 router.get("/:id", async (req, res) => {
   try {
     const post = await Post.findById(req.params.id);
+    if (!post) {
+      return res.status(404).json("Post not found");
+    }
     res.status(200).json(post);
   } catch (err) {
     res.status(500).json(err);
@@ -68,6 +77,11 @@ router.get("/timeline/all", async (req, res) => {
 //get all post for timeline
 router.get("/timeline/all/:id", async (req, res) => {
   try {
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      return res.status(404).json("User not found!");
+    }
+
     const userFollowing = await User.findById(req.params.id);
     userFollowingFollowing = userFollowing.following;
     userFollowingFollowing.push(req.params.id);
@@ -84,6 +98,10 @@ router.get("/timeline/all/:id", async (req, res) => {
 //get all post by id user
 router.get("/all/:id", async (req, res) => {
   try {
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      return res.status(404).json("User not found!");
+    }
     const userPost = await Post.find({ userId: req.params.id }).sort({
       createdAt: "desc",
     });
@@ -93,20 +111,8 @@ router.get("/all/:id", async (req, res) => {
   }
 });
 
-// get last post by id user
-router.get("/last/:id", async (req, res) => {
-  try {
-    const userPost = await Post.findOne({ userId: req.params.id }).sort({
-      createdAt: "desc",
-    });
-    res.json(userPost);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
 // search posts by query
-router.get("/search/:search/", async (req, res) => {
+router.get("/search/:search", async (req, res) => {
   try {
     var posts = await Post.find({
       desc: { $regex: req.params.search, $options: "i" },
