+++ routes/projects.js
@@ -1,11 +1,11 @@
 const router = require("express").Router();
 const Project = require("../models/Project");
-const Album = require("../models/Album");
+const User = require("../models/User");
 
 //create project
 router.post("/", async (req, res) => {
-  const newProject = new Project(req.body);
   try {
+    const newProject = new Project(req.body);
     const savedProject = await newProject.save();
     res.status(200).json(savedProject);
   } catch (err) {
@@ -15,13 +15,19 @@ router.post("/", async (req, res) => {
 
 //update project
 router.put("/:id", async (req, res) => {
-  const project = await Project.findById(req.params.id);
   try {
-    if (project.userId === req.body.userId) {
+    const project = await Project.findById(req.params.id);
+    if (!project) {
+      return res.status(404).json("Project not found");
+    }
+    if (
+      project.userId === req.body.userId ||
+      project.participants.includes(req.body.userId)
+    ) {
       await project.updateOne({ $set: req.body });
       res.status(200).json("Project updated!");
     } else {
-      res.status(500).json("You can only update your project");
+      res.status(403).json("You can only update your project");
     }
   } catch (err) {
     res.status(500).json(err);
@@ -30,15 +36,18 @@ router.put("/:id", async (req, res) => {
 
 //delete project
 router.delete("/:id", async (req, res) => {
-  const project = await Project.findById(req.params.id);
   try {
+    const project = await Project.findById(req.params.id);
+    if (!project) {
+      return res.status(404).json("Project not found");
+    }
     if (project.userId === req.body.userId) {
       // await project.deleteOne({ $set: req.body });
       await Project.findByIdAndDelete(req.params.id);
 
       res.status(200).json("Project deleted!");
     } else {
-      res.status(500).json("You can only delete your project");
+      res.status(403).json("You can only delete your project");
     }
   } catch (err) {
     res.status(500).json(err);
@@ -49,17 +58,10 @@ router.delete("/:id", async (req, res) => {
 router.get("/:id", async (req, res) => {
   try {
     const project = await Project.findById(req.params.id);
+    if (!project) {
+      return res.status(404).json("Project not found");
+    }
     res.status(200).json(project);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-//get all project
-router.get("/timeline/all", async (req, res) => {
-  try {
-    const userProject = await Project.find().sort({ createdAt: "desc" });
-    res.json(userProject);
   } catch (err) {
     res.status(500).json(err);
   }
@@ -68,7 +70,13 @@ router.get("/timeline/all", async (req, res) => {
 //get all project by id user
 router.get("/all/:id", async (req, res) => {
   try {
-    const userProject = await Project.find({ userId: req.params.id }).sort({
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      return res.status(404).json("User not found!");
+    }
+    const userProject = await Project.find({
+      $or: [{ userId: req.params.id }, { participants: req.params.id }],
+    }).sort({
       createdAt: "desc",
     });
     res.json(userProject);
@@ -77,20 +85,8 @@ router.get("/all/:id", async (req, res) => {
   }
 });
 
-// get last project by id user
-router.get("/last/:id", async (req, res) => {
-  try {
-    const userProject = await Project.findOne({ userId: req.params.id }).sort({
-      startDate: "desc",
-    });
-    res.json(userProject);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
 // search projects by query
-router.get("/search/:search/", async (req, res) => {
+router.get("/search/:search", async (req, res) => {
   try {
     var projects = await Project.find({
       $or: [
@@ -102,6 +98,93 @@ router.get("/search/:search/", async (req, res) => {
     });
 
     res.status(200).json(projects);
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+// project suggestion
+router.get("/suggest/:id", async (req, res) => {
+  try {
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      return res.status(404).json("User not found!");
+    }
+    const userProject = await Project.find({
+      userId: { $ne: req.params.id },
+      participants: { $ne: req.params.id },
+      isPublic: true,
+      endDate: null,
+    })
+      .sort({
+        createdAt: "desc",
+      })
+      .limit(3);
+    res.json(userProject);
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+// request participation
+router.put("/request/:id", async (req, res) => {
+  try {
+    const project = await Project.findById(req.params.id);
+    if (!project) {
+      return res.status(404).json("Project not found");
+    }
+
+    // tambah request
+    let projectRequest = project.requests;
+    projectRequest.push(req.body.userId);
+    await project.updateOne({
+      $set: { requests: projectRequest },
+    });
+    res.status(200).json("Project requested!");
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+// cancel request participation
+router.put("/cancelrequest/:id", async (req, res) => {
+  try {
+    const project = await Project.findById(req.params.id);
+    if (!project) {
+      return res.status(404).json("Project not found");
+    }
+
+    // hapus request
+    let projectRequest = project.requests;
+    projectRequest = projectRequest.filter((e) => e !== req.body.userId);
+    await project.updateOne({
+      $set: { requests: projectRequest },
+    });
+    res.status(200).json("Project requested!");
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+// accept request participation
+router.put("/accept/:id", async (req, res) => {
+  try {
+    const project = await Project.findById(req.params.id);
+    if (!project) {
+      return res.status(404).json("Project not found");
+    }
+
+    // hapus request
+    let projectRequest = project.requests;
+    projectRequest = projectRequest.filter((e) => e !== req.body.userId);
+
+    // tambah participant
+    let projectParticipants = project.participants;
+    projectParticipants.push(req.body.userId);
+    await project.updateOne({
+      $set: { participants: projectParticipants, requests: projectRequest },
+    });
+    res.status(200).json("Participant added!");
   } catch (err) {
     res.status(500).json(err);
   }
